MailAPI/requirements.txt
# MailAPI Dependencies

# Email processing
# Note: imaplib and email are built-in Python libraries

# Document processing
pypdf>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0

# HTML to text conversion
html2text>=2020.1.16

# Date and time parsing
python-dateutil>=2.8.0
pytz>=2024.1

# Configuration management
PyYAML>=6.0
python-dotenv>=1.0.0

# Optional: OCR support (uncomment if needed)
# pytesseract>=0.3.10

# Optional: Advanced NLP (uncomment if needed)
# nltk>=3.8.1
# spacy>=3.7.0
